requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
opencv-python==4.5.3.56
numpy==1.21.2
